feat: Added the `textStyleBuilder` parameter to the `TagTextEditingController`
Added the `textStyleBuilder` parameter to the `TagTextEditingController` constructor for determining the style of tags in the TextField, in favor of the `textStyle` parameter of the `TagStyle` class, which has been deprecated. The builder allows you to use the TextField's `BuildContext` when initializing the `TagTextEditingController` object in the `initState` method of a StatefulWidget, such that you can access inherited styles such as `Theme.of(context).textTheme`.